Split IFhirStore into 2 interfaces: IFhirStore and IAsyncFhirStore
Mark IFhirStore as Obsolete.

One change was made that could possibly break backward compatibility: MongoStoreBuilder was removed, as it seems it's unused anyway, and I found no good way to leave it and to support it.
