+++ src/Spark.Engine/Store/Interfaces/IFhirStore.cs
@@ -7,21 +7,20 @@
  * available at https://raw.githubusercontent.com/FirelyTeam/spark/stu3/master/LICENSE
  */
 
+using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Spark.Engine.Core;
 
 namespace Spark.Engine.Store.Interfaces
 {
+    [Obsolete("Use IAsyncFhirStore instead")]
     public interface IFhirStore
     {
         void Add(Entry entry);
-        Task AddAsync(Entry entry);
 
         Entry Get(IKey key);
-        Task<Entry> GetAsync(IKey key);
 
         IList<Entry> Get(IEnumerable<IKey> localIdentifiers);
-        Task<IList<Entry>> GetAsync(IEnumerable<IKey> localIdentifiers);
     }
 }
